chore: Move net_gen into devices::virtio::net
In our ongoing effort to merge our many crates into one, tackle net_gen
as a free-standing crate without dependencies and only one dependent.
