+++ src/utils/Cargo.toml
@@ -19,7 +19,5 @@ versionize_derive = "0.1.6"
 vmm-sys-util = "0.11.2"
 vm-memory = { version = "0.12.0", features = ["backend-mmap", "backend-bitmap"] }
 
-net_gen = { path = "../net_gen" }
-
 [dev-dependencies]
 serde_json = "1.0.99"
