+++ src/vmm/Cargo.toml
@@ -36,7 +36,6 @@ base64 = "0.13.0"
 bincode = "1.2.1"
 micro_http = { git = "https://github.com/firecracker-microvm/micro-http" }
 
-net_gen = { path = "../net_gen" }
 seccompiler = { path = "../seccompiler" }
 snapshot = { path = "../snapshot"}
 utils = { path = "../utils" }
