+++ src/vmm/src/devices/virtio/net/test_utils.rs
@@ -262,7 +262,7 @@ pub fn if_index(tap: &Tap) -> i32 {
     let sock = create_socket();
     let ifreq = IfReqBuilder::new()
         .if_name(&tap.if_name)
-        .execute(&sock, c_ulong::from(net_gen::sockios::SIOCGIFINDEX))
+        .execute(&sock, c_ulong::from(super::gen::sockios::SIOCGIFINDEX))
         .unwrap();
 
     // SAFETY: Using this union variant is safe since `SIOCGIFINDEX` returns an integer.
@@ -285,13 +285,13 @@ pub fn enable(tap: &Tap) {
     IfReqBuilder::new()
         .if_name(&tap.if_name)
         .flags(
-            (net_gen::net_device_flags_IFF_UP
-                | net_gen::net_device_flags_IFF_RUNNING
-                | net_gen::net_device_flags_IFF_NOARP)
+            (crate::devices::virtio::net::gen::net_device_flags_IFF_UP
+                | crate::devices::virtio::net::gen::net_device_flags_IFF_RUNNING
+                | crate::devices::virtio::net::gen::net_device_flags_IFF_NOARP)
                 .try_into()
                 .unwrap(),
         )
-        .execute(&sock, c_ulong::from(net_gen::sockios::SIOCSIFFLAGS))
+        .execute(&sock, c_ulong::from(super::gen::sockios::SIOCSIFFLAGS))
         .unwrap();
 }
 
@@ -351,11 +351,11 @@ pub mod test {
     use std::{cmp, fmt, mem};
 
     use event_manager::{EventManager, SubscriberId, SubscriberOps};
-    use net_gen::ETH_HLEN;
     use utils::vm_memory::{Address, Bytes, GuestAddress, GuestMemoryMmap};
 
     use crate::check_metric_after_block;
     use crate::devices::virtio::net::device::vnet_hdr_len;
+    use crate::devices::virtio::net::gen::ETH_HLEN;
     use crate::devices::virtio::net::test_utils::{
         assign_queues, default_net, inject_tap_tx_frame, NetEvent, NetQueue, ReadTapMock,
     };
